Repair post-rebase module layout (#551)
Preserve the intended manifest-query tests while repairing replayed fixtures.
Split the AST value type and rendering tests into focused modules so the
rebased branch satisfies the repository's module-size policy without changing
the public AST path.
